[ENT-1878] - watermarking in share link index
To support some of the work we need to do for session based watermarking
we need to be able to pass data from a share link to an asset's
serializer so we can see if the share link it's loaded from has
watermarking enabled.
This adds a assigns/2 macro to the serializer that allow specifying
data you want to assign to the conn for every relationship loaded under
a certain entity.