[SERV-1854] Safe atom conversion
There are a few places where we convert strings into atoms for use
in including/excluding certain attributes in the serialized maps we
return. These strings come from the client, meaning they are arbitrary
and could be of any form.

At the moment, we use `String.to_atom/1` to convert these values into
atoms. This is unsafe for a number of reasons. For one, that function
will raise if a too-long string is passed in. But more importantly, it
does not guard against hitting the BEAM atom limit.

Going forward, we will use `String.to_existing_atom/1` for converting
this arbitrary field filter strings into atoms. As the name suggests,
the function will raise an exception if the string-as-atom hasn't been
previously defined. This should be okay for our use cases, since all
attributes should be defined with an atom in a serializer prior to
serialization. If we cannot convert the string, we simply return `nil`.